Implement kqueue multiplexer backend for macos
