+++ libnet/net/fwd.hpp
@@ -16,8 +16,8 @@ namespace net {
 // -- classes ------------------------------------------------------------------
 
 class acceptor;
-class epoll_multiplexer;
 class multiplexer;
+class multiplexer_impl;
 class pollset_updater;
 class socket_manager;
 class socket_manager_factory;
